+++ mods/BrewLAN_Units/BrewAir/mod_info.lua
@@ -1,7 +1,7 @@
 name = "BrewLAN: Penetration"
-uid = "25D57D85-9JA7-D842-BREW-AIR00000000021"
-version = 2.1
-copyright = "© 2018–2020 Sean Wheeldon"
+uid = "25D57D85-9JA7-D842-BREW-AIR00000000022"
+version = 2.2
+copyright = "© 2018–2022 Sean Wheeldon"
 description = "Adds extended aircraft options."
 author = "Balthazar"
 icon = "/mods/BrewLAN_Units/BrewAir/mod_icon.png"