+++ mods/BrewLAN_Units/BrewAir/projectiles/CIFNeutronCarpetBomb/CIFNeutronCarpetBomb_script.lua
@@ -1,0 +1,6 @@
+
+CIFNeutronClusterBomb01 = Class(import('/lua/cybranprojectiles.lua').CNeutronClusterBombProjectile) {
+    PolyTrail = '/effects/emitters/cybran_iridium_missile_polytrail_01_emit.bp',
+}
+
+TypeClass = CIFNeutronClusterBomb01