+++ mods/BrewLAN_Units/BrewAir/units/SRA0308/SRA0308_script.lua
@@ -1,0 +1,14 @@
+local CAirUnit = import('/lua/cybranunits.lua').CAirUnit
+local CIFBombNeutronWeapon = import('/lua/cybranweapons.lua').CIFBombNeutronWeapon
+
+SRA0308 = Class(CAirUnit) {
+    Weapons = {
+        Bomb = Class(CIFBombNeutronWeapon) {},
+    },
+
+    OnStopBeingBuilt = function(self,builder,layer)
+        CAirUnit.OnStopBeingBuilt(self,builder,layer)
+        self:SetScriptBit('RULEUTC_StealthToggle', true)
+    end,
+}
+TypeClass = SRA0308